// src/types/bookmarks.d.ts
export interface BookmarkData {
  title: string;
  creator: string;
  url: string;
  tags: string[];
}

export interface ArticleData {
  title: string;
  author: string;
  site: string;
  url: string;
  tags: string[];
}

export type PageData = ArticleData | BookmarkData;

export interface RedditData {
  title: string;
  content: string;
  subreddit: string;
  url: string;
  tags: string[];
}

export interface TwitterData {
  tweet: string;
  user: string;
  url: string;
}

export interface ParsingService {
  title: RegExp;
  creator: RegExp;
  url: RegExp;
  [key: string]: RegExp;
}

export interface ParsingPatterns {
  castro: ParsingService;
  overcast: ParsingService;
  title: RegExp[];
  [key: string]: ParsingService | RegExp[];
}
// ...
export interface BookmarkingResponse {
  success: boolean;
  message: string;
  source: string;
}